Updated for new conda env approach
In this version, user creates conda envs that are used within shell commands inside rules.

This is not the best approach - better would be to specify the env file in the relevant rule and have Snakemake create the env internally, but there's issues with downloading the packages when I do it that way. I am still attempting to resolve that so it works.