Add DANPOS python scripts
Added DANPOS python scripts.

The DANPOS3 version on Petrichor has a known issue (https://www.biostars.org/p/457084/) using a misnamed function in the `reads.py` package. I have corrected this issue and included the python scripts in this repo to ensure that others are using the corrected version of DANPOS3.

Added documentation for why DANPOS3 is included.